srvtools/pyinfra/requirements.txt
black >= 23.3.0, < 24
boto3 >= 1.26.142, < 2
mypy >= 1.3.0, < 2
passlib >= 1.7.4, < 2
pyinfra >= 2.6.1, < 3
PyYAML >= 6.0, < 7
types-passlib
types-PyYAML
